fix: page number error

See merge request s10-bigdata-recom-sub2/S10P22A705!156

+++ backend/src/main/java/com/jshi/laughtale/chapter/controller/ChapterController.java
@@ -44,7 +44,7 @@ public class ChapterController {
 	public ResponseEntity<?> getChapterPagination(
 		@PathVariable Long chapterId
 	){
-		List<Integer> pagination = chapterService.getChapterPagination(chapterId);
+		List<Long> pagination = chapterService.getChapterPagination(chapterId);
 
 		return ResponseEntity.ok(ChapterPaginationDto
 			.Response
